requirements.txt
mysql-connector-python==8.1.0
protobuf==4.21.12
python-dotenv==1.0.0
